requirements.txt
python-dotenv
langgraph
langchain-core
langchain-google-genai
google-generativeai
python-dotenv
typing-extensions
ipython
